*: decode iat and exp as int64
Since #56 this package encodes the 'iat' and 'exp' claims as int64
instead of float64. However it still decodes them as a float64.
This causes comparisions to fail in tests that use some version of
reflect.DeepEqual. See coreos/dex#418.

To avoid this, attempt to decode the 'iat' and 'exp' explicitly as
int64 if possible.