*: remove dependency on capnslog
capnslog hijacks the standard loggers as an import side effect. For
some reason in dex this is causing the standard logger to not log
anything. For example net/http just swallows panics in handlers
without printing stack traces.

I'm having trouble tracking this down in dex, but removing the
capnslog's logging hijacker init() fixed this, so we're just going
to remove the dependency altogether. Importing go-oidc shouldn't
alter the standard logger anyway.