jose: remove hmac signing and verification code
This was brought up during disucssions about #98. The HMAC logic
isn't used by this package or dex. We don't support client_secret_jwt
client authentication[0] in dex and don't provide an interface for it
in our client package.

Remove the HMAC code for now. We don't want to encourage its use and
can revert this change if ever need to support it.

This change has been tested with the dex repo and passes those unit
tests locally.

[0] https://openid.net/specs/openid-connect-core-1_0.html#ClientAuthentication