oauth2: don't send scopes to token endpoint
Not specifying a "scope" parameter for a refresh token means using
the original scopes. However, it's currently impossible to redeem a
refresh token without specify these. This is very problematic when
a client doesn't know the original scopes.

Rework the parameters requested of the token endpoint to be more in
line with the spec. Also, never specifiy the scope parameter when
refreshing a token which is what golang.org/x/oauth2 does[0].

The behavior we care about is:

* Refresh tokens don't require "scope" and doesn't use "redirect_uri"[1]
* Access token requests don't use "scope"[2]

[0] https://github.com/golang/oauth2/blob/1e695b1c8febf17aad3bfa7bf0a819ef94b98ad5/oauth2.go#L230-L233
[1] https://tools.ietf.org/html/rfc6749#section-6
[2] https://tools.ietf.org/html/rfc6749#section-4.1.3