special case google accounts issuer validation
Google sometimes returns "accounts.google.com" instead of
"https://accounts.google.com". Special case this.

Originally we merged a new verification flag, but since this should
only be used with Google, and should always be on with Google, it
makes more sense for us to detect this case transparently.

Fixes #125