jwks: Don't store context in remoteKeySet
The docs for the context package say:

    Do not store Contexts inside a struct type; instead, pass a Context
    explicitly to each function that needs it.

Previously the context passed to NewProvider would get stashed away in
the remoteKeySet structure, then maybe used later. If the context that
created the provider had been cancelled, keys could not be fetched.

For a use case where this is bad, consider the user of this library is
an HTTP server. It has a REST endpoint which can configure OIDC
providers to use for authentication. It then has other endpoints that
use those providers. When the configure endpoint is called, it creates a
context and calls NewProvider(). When the handler for the configure
endpoint is done, it cancels the context.

When the provider returned by NewProvider() is subsequently used, keys
can not be verified: they fail with "context canceled".

Fortunately the keys get updated from Verify(), which already takes a
context. So we can just use that. This does mean if the context is being
used to change the default HTTP client it's necessary to do it on the
context passed to Verify() as well as NewProvider().