oidc: Allows client to read configuration
This provides a copy of the running configuration to clients after it
has been obtained from a Discovery Endpoint.  This is accomplished by
adding a Config() method to the oidc.Provider type, which loads a new
oidc.ProviderConfig struct with the values the Provider is running with
and returns it to the user.