oidc: verify the ID Token's signature before processing claims
This change updates the verification logic of this library to first
validate the ID Token instead of parsing claims. This hopefully makes it
harder for a malicious client to provide an invalid token for validation
that's crafted to cause this package to over-allocate memory. See the
associated bug and CVE-2025-27144.

Fixes https://github.com/coreos/go-oidc/pull/463